requirements.txt
albumentations>=1.4.3
coremltools>=7.1
ipython>=8.12.3
matplotlib>=3.8.4
numpy>=1.25.1
onnx>=1.16.0
onnxruntime>=1.17.3
onnxsim>=0.4.36
opencv_python>=4.8.0.74
openpyxl>=3.1.2
pandas>=2.2.2
Pillow>=10.3.0
psutil>=5.9.8
PyYAML>=6.0.1
Requests>=2.31.0
scipy>=1.13.0
seaborn>=0.13.2
torch>=2.2.2
torchvision>=0.17.2
tqdm>=4.66.2
pyzbar>=0.1.9
